rag-mcp-server/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb==0.5.23
sentence-transformers==3.3.1
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.1
smbprotocol==1.14.0
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
markdown==3.7
beautifulsoup4==4.12.3
chardet==5.2.0
watchdog==6.0.0
aiofiles==24.1.0
httpx==0.28.1
